+++ src/main/java/com/swmStrong/demo/domain/user/facade/UserUpdateProviderImpl.java
@@ -2,7 +2,7 @@ package com.swmStrong.demo.domain.user.facade;
 
 import com.swmStrong.demo.common.exception.ApiException;
 import com.swmStrong.demo.common.exception.code.ErrorCode;
-import com.swmStrong.demo.domain.global.Role;
+import com.swmStrong.demo.domain.common.enums.Role;
 import com.swmStrong.demo.domain.user.entity.User;
 import com.swmStrong.demo.domain.user.repository.UserRepository;
 import org.springframework.stereotype.Service;
